% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings2_doc.r
\docType{data}
\name{ratings2}
\alias{ratings2}
\title{Dataset of ratings that 4 raters assigned to 15 subjects}
\format{
Each row of this dataset describes the ratings that a subject 
received from the 4 raters.
 \describe{
    \item{RaterA}{Ratings of 15 subjects by Rater A}
    \item{RaterB}{Ratings of 15 subjects by Rater B}
    \item{RaterC}{Ratings of 15 subjects by Rater C}
    \item{RaterD}{Ratings of 15 subjects by Rater D}
 }
}
\source{
This is a fictitious dataset used for testing the difference between
        2 agreement coefficients for statistical significance.
}
\usage{
ratings2
}
\description{
This dataset contains ratings on 15 subjects from 4 raters named RaterA,
RaterB, RaterC and RaterD
}
\keyword{datasets}
