% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.ac2.r
\name{ttest.ac2}
\alias{ttest.ac2}
\title{Paired t-test for the difference of 2 AC2 coefficients.}
\usage{
ttest.ac2(
  g1.ratings,
  g2.ratings,
  weights = "unweighted",
  conflev = 0.95,
  N = Inf
)
}
\arguments{
\item{g1.ratings}{is a mandatory parameter representing the first data frame
of ratings.}

\item{g2.ratings}{is a mandatory parameter representing the second data frame
of ratings.}

\item{weights}{is an optional parameter that defines the weights needed in a
weighted analysis. It's default value is ``unweighted'' which requests the
unweighted analysis.}

\item{conflev, }{is an optional parameter representing the confidence level.
It's default value is 0.95.}

\item{N}{is an optional parameter representing the size of the subject
population. It's default value is infinity.}
}
\description{
The ttest.ac2 function performs the paired t-test for testing
the difference between two correlated Gwet's \eqn{AC_2} coefficients for
statistical significance. It implements the linearization method of
Gwet (2016).
}
\details{
The 2 required parameters "g1.ratings" and "g2.ratings",
which are the 2 groups of raters being compared, must have the
exact same number of rows, and each column represents one rater and
contains its ratings (numeric or alphabetic). All subjects that are not
rated by any rater must be excluded from the dataset.
}
\references{
Gwet, K. L. (2008). Computing inter-rater reliability and its variance in
the presence of high agreement. \emph{British Journal of Mathematical and
Statistical Psychology}, 61, 29-48.

Gwet, K. L. (2016). Testing the Difference of Correlated Agreement
Coefficients for Statistical Significance, \emph{Educational and
Psychological Measurement}, Vol 76(4) 609-637.
}
