% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.pa.r
\name{ttest.pa}
\alias{ttest.pa}
\title{Paired t-test for the difference of 2 percent agreement coefficients}
\usage{
ttest.pa(
  g1.ratings,
  g2.ratings,
  weights = "unweighted",
  conflev = 0.95,
  N = Inf
)
}
\arguments{
\item{g1.ratings}{is a mandatory parameter representing the first data frame
of ratings.}

\item{g2.ratings}{is a mandatory parameter representing the second data frame
of ratings.}

\item{weights}{is an optional parameter that defines the weights needed in a
weighted analysis. It's default value is ``unweighted'' which requests the
unweighted analysis.}

\item{conflev, }{is an optional parameter representing the confidence level.
It's default value is 0.95.}

\item{N}{is an optional parameter representing the size of the subject
population. It's default value is infinity.}
}
\description{
This function performs the paired t-test for testing the difference between
two correlated percent agreement coefficients for statistical significance.
It implements the linearization method of Gwet (2016).
}
\details{
This function takes 2 required parameters, which are the 2 groups of raters
being compared: "g1.ratings" and "g2.ratings." Both datasets must have the
exact same number of rows, and each column represents one rater and
contains its ratings (numeric or alphabetic).
The user must exclude all subjects that are not rated by any rater.
}
\references{
Gwet, K.L. (2016). ``Testing the Difference of Correlated Agreement
Coefficients for Statistical Significance,'' \emph{Educational and
Psychological Measurement}, \strong{76}(4) 609-637.
}
