% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.matrix.r
\name{weight.mat.fn}
\alias{weight.mat.fn}
\title{Create a weight matrix based raw data and weight set name}
\usage{
weight.mat.fn(ratings, weights = "unweighted")
}
\arguments{
\item{ratings}{is a mandatory parameter representing the input data frame
of raw ratings.}

\item{weights}{is an optional parameter that defines the weights needed in a
weighted analysis. It's default value is ``unweighted'' which requests the
unweighted analysis.}
}
\description{
The weight.mat.fn function produces the weight matrix needed
to compute weighted agreement coefficients for categorical ratings.
}
\references{
Gwet, K. L. (2021). \emph{Handbook of Inter-Rater Reliability - Volume 1: Analysis
of Categorical Ratings}. AgreeStat Analytics.
}
